+++ ChangeLog.md
@@ -4,15 +4,15 @@
 * No longer depend on env.
 * Switched from ore-tasks to rubygems-tasks ~> 0.1.
 * Added {Ore::Naming} from `ore-core`.
+* Added {Ore::Options}.
 * Added {Ore::Actions}.
 * Added {Ore::Template::Helpers#rubygems_tasks?}.
 * Added {Ore::Template::Helpers#bundler_tasks?}.
 * Added {Ore::Template::Helpers#gem_package_task?}.
 * Added {Ore::Template::Directory#dependencies}.
 * Added {Ore::Template::Directory#development_dependencies}.
-* Added {Ore::Generator::OPTIONS}.
 * Renamed `Ore::Config.default_options` to {Ore::Config.options}.
-* Renamed `Ore::Generator.defaults` to {Ore::Generator.options}.
+* Renamed `Ore::Generator.defaults` to {Ore::Options.defaults}.
 * Renamed `Ore::Generator.templates` to {Ore::Template.templates}.
 * Renamed `Ore::Generator.template?` to {Ore::Template.template?}.
 * Renamed `Ore::Generator.register_template` to {Ore::Template.register}.
@@ -122,7 +122,7 @@
 * Added {Ore::Config.disable!}.
 * Added `Ore::Config.default_options`.
 * Added `Ore::Generator.defaults`.
-* Added {Ore::Generator.generator_option}.
+* Added `Ore::Generator.generator_option`.
 * Added `vendor/cache/*.gem` to `.gitignore` if `--bundler` is specified.
 * Attempt to auto-load `ore/specification` in the generated `*.gemspec`
   files.